+++ src/index.js
@@ -9,10 +9,7 @@ import {
 	enrollUserInGuidedCourse,
 	fetchEnrollmentPageMetadata,
 	guidedCourses,
-	pinGuidedCourse,
-	pinnedGuidedCourses,
-	unEnrollUserInGuidedCourse,
-	unPinGuidedCourse
+	unEnrollUserInGuidedCourse
 } from './services/content-org/guided-courses.ts';
 
 import {
@@ -59,6 +56,7 @@ import {
 import {
 	contentStatusCompleted,
 	contentStatusReset,
+	contentStatusStarted,
 	getAllCompleted,
 	getAllStarted,
 	getAllStartedOrCompleted,
@@ -95,8 +93,14 @@ import {
 } from './services/forum.js';
 
 import {
-	fetchAwardsForUser
-} from './services/gamification/awards.js';
+	createForumCategory,
+	fetchForumCategories
+} from './services/forums/categories.ts';
+
+import {
+	fetchAwardsForUser,
+	fetchCertificate
+} from './services/gamification/awards.ts';
 
 import {
 	applyCloudflareWrapper,
@@ -148,6 +152,7 @@ import {
 	postContentLiked,
 	postContentReset,
 	postContentRestore,
+	postContentStart,
 	postContentUnliked,
 	postPlaylistContentEngaged,
 	postRecordWatchSession,
@@ -219,6 +224,8 @@ import {
 } from './services/sanity.js';
 
 import {
+	confirmEmailChange,
+	requestEmailChange,
 	resetPassword,
 	sendAccountSetupEmail,
 	sendPasswordResetEmail,
@@ -252,6 +259,11 @@ import {
 } from './services/user/management.js';
 
 import {
+	fetchMemberships,
+	fetchRechargeTokens
+} from './services/user/memberships.js';
+
+import {
 	deleteNotification,
 	fetchLiveEventPollingState,
 	fetchNotificationSettings,
@@ -265,6 +277,10 @@ import {
 	startLiveEventPolling,
 	updateNotificationSetting
 } from './services/user/notifications.js';
+
+import {
+	fetchCustomerPayments
+} from './services/user/payments.ts';
 
 import {
 	fetchUserPermissions,
@@ -308,6 +324,10 @@ import {
 	updateUserPractice
 } from './services/userActivity.js';
 
+import {
+	 default as EventsAPI 
+} from './services/eventsAPI';
+
 export {
 	addContextToContent,
 	addItemToPlaylist,
@@ -319,10 +339,13 @@ export {
 	buildImageSRC,
 	calculateLongestStreaks,
 	closeComment,
+	confirmEmailChange,
 	contentStatusCompleted,
 	contentStatusReset,
+	contentStatusStarted,
 	convertToTimeZone,
 	createComment,
+	createForumCategory,
 	createPlaylist,
 	createPracticeNotes,
 	deleteComment,
@@ -348,6 +371,7 @@ export {
 	fetchByRailContentIds,
 	fetchByReference,
 	fetchCarouselCardData,
+	fetchCertificate,
 	fetchChatAndLiveEnvent,
 	fetchChatSettings,
 	fetchCoachLessons,
@@ -362,7 +386,9 @@ export {
 	fetchContentPageUserData,
 	fetchContentProgress,
 	fetchContentRows,
+	fetchCustomerPayments,
 	fetchEnrollmentPageMetadata,
+	fetchForumCategories,
 	fetchFoundation,
 	fetchGenreLessons,
 	fetchHandler,
@@ -375,6 +401,7 @@ export {
 	fetchLikeCount,
 	fetchLiveEvent,
 	fetchLiveEventPollingState,
+	fetchMemberships,
 	fetchMetadata,
 	fetchMethod,
 	fetchMethodChildren,
@@ -395,6 +422,7 @@ export {
 	fetchRecent,
 	fetchRecentActivitiesActiveTabs,
 	fetchRecentUserActivities,
+	fetchRechargeTokens,
 	fetchRelatedLessons,
 	fetchRelatedRecommendedContent,
 	fetchRelatedSongs,
@@ -486,13 +514,12 @@ export {
 	openComment,
 	otherStats,
 	pauseLiveEventPolling,
-	pinGuidedCourse,
 	pinProgressRow,
-	pinnedGuidedCourses,
 	postContentComplete,
 	postContentLiked,
 	postContentReset,
 	postContentRestore,
+	postContentStart,
 	postContentUnliked,
 	postPlaylistContentEngaged,
 	postRecordWatchSession,
@@ -508,6 +535,7 @@ export {
 	replyToComment,
 	reportComment,
 	reportPlaylist,
+	requestEmailChange,
 	reset,
 	resetPassword,
 	restoreComment,
@@ -525,7 +553,6 @@ export {
 	toDayjs,
 	togglePlaylistPrivate,
 	unEnrollUserInGuidedCourse,
-	unPinGuidedCourse,
 	unassignModeratorToComment,
 	unblockUser,
 	undeletePlaylist,
@@ -543,3 +570,5 @@ export {
 	verifyImageSRC,
 	verifyLocalDataContext,
 };
+
+export default EventsAPI